OnNavigatedTo for device GoBack
