+++ Source/Xamarin/Prism.Forms/Behaviors/MultiPageActiveAwareBehavior{T}.cs
@@ -1,4 +1,5 @@
-﻿using Prism.Navigation;
+﻿using Prism.Common;
+using Prism.Navigation;
 using System;
 using Xamarin.Forms;
